project/microservices/sentiment/requirements.txt
faust-streaming
faust-streaming-rocksdb
transformers[torch]
